refactor: 이름 잘못된 테스트클래스 수정 및 readMe 수정
