+++ STEP4README.md
@@ -11,7 +11,7 @@
       - [x] ,로 구분하고, 구분한 개수만큼 자동차 개수로 값 저장
     - 게임 완료 후 우승자 출력.
       - [x] 제일 거리가 긴 Car 객체를 찾아서 저장후 반환한다.
-      - [ ] 우승자를 출력하는 메서드를 ResultView에 추가
+      - [x] 우승자를 출력하는 메서드를 ResultView에 추가
 
 
 #구조
