+++ src/test/kotlin/racingcar/CarsTest.kt
@@ -12,6 +12,14 @@ class CarsTest {
         val testCarList = listOf(Car("red",1),Car("blue",1))
         val testCars = Cars(testCarList)
         assertThat(cars).isEqualTo(testCars)
+    }
 
+    @Test
+    fun createWinningCarListTest()  {
+        val testCarList = listOf(Car("red",2),Car("blue",2),
+            Car("kaki",1))
+        val winningList = listOf(Car("red",2),Car("blue",2))
+        val cars = Cars(testCarList)
+        assertThat(cars.findWinner()).isEqualTo(winningList)
     }
 }
